Split multipwt.py into multiple scripts (issue #25).
