+++ mpwt/__init__.py
@@ -1,2 +1,3 @@
-from mpwt.multipwt import multiprocess_pwt
+from mpwt.pwt_wrapper import run_pwt, run_pwt_dat
+from mpwt.mpwt_workflow import multiprocess_pwt
 from mpwt.utils import cleaning, cleaning_input, find_ptools_path, list_pgdb, remove_pgbds
