+++ mpwt/utils.py
@@ -144,3 +144,19 @@ def cleaning_input(input_folder, verbose=None):
             if verbose:
                 species = genbank_path.split('/')[-2]
                 logger.info('Remove ' + species + ' temporary datas.')
+
+
+def permission_change(folder_pathname):
+    """
+    Give permission to output files inside a folder.
+    Used for log files and PGDB/dat files.
+
+    Args:
+        folder_pathname (str): pathname to the folder which permissions will be changed
+    """
+    os.chmod(folder_pathname, 0o777)
+    for root, subfolders, subfiles in os.walk(folder_pathname):
+        for subfolder in subfolders:
+            os.chmod(os.path.join(root, subfolder), 0o777)
+        for subfile in subfiles:
+            os.chmod(os.path.join(root, subfile), 0o777)
