+++ test/test/fatty_acid_beta_oxydation_I/fatty_acid_beta_oxydation_I.fasta
@@ -1,4 +1,4 @@
->fatty_acid_beta_oxydation_I
+>NC_000913_3
 TTACGCGGCTTCAACTTTCCGCACTTTCTCCGGCAACTTTACCGGCTTCGTCGCCAGCTC
 TTCCGGATCAAAGTCATCAACGTTAATACTGCGCAGACGGCTTTCTTCAGCTTTCACCAG
 AATAGCGGCTTCATCTTTATCAATCAGCCCCTTCACCAGCGCGTTGTGCGCCAGTTCATC
