+++ test/test/fatty_acid_beta_oxydation_I_pf/NC_000913_3.fasta
