+++ test/test/fatty_acid_beta_oxydation_I_pf/NC_000913_3.pf
@@ -1,5 +1,5 @@
 ;;;;;;;;;;;;;;;;;;;;;;;;;
-;; fatty_acid_beta_oxydation_I_pf
+;; NC_000913_3
 ;;;;;;;;;;;;;;;;;;;;;;;;;
 ID	gene226
 NAME	gene226
